Revert "Add support for parsing parenthesis in URLS ()"
This reverts commit 25bad7cecd31c6a2fe8adadaf123d4b98e53d08d.

This was in fact a extremely bad decision.
The best way to proceed is to encode special characters whenever
possible.

The critical effect of this was found when parsing URLs inside CCS
files where it's possible to see `url(../path/to/file.jpg')` and this
change has caused confusion, making everything, a big URL, including the
bits outside the parenthesis.

Reverting this for now. 💆‍♂️