Pause on source-map implementation
After hours researching and trying to implement source maps, it seems
currently it's not possible to have multiple sections of a document
pointing to different mappings due to the nature of mappings always
using the base reference on the minified document starting with the
column index, meaning it would always be a single line.
```
[mincol,sources_index,rawline,rawcol,names_index]
```
considering this, mincol will not accept muliple lines.
The v3 spec for source maps has a proposal for the `sections` see
https://sourcemaps.info/spec.html where a user would be able to declare
multiple sections for different parts of a document, however this hasn't
been implemented yet and I am not sure it ever will be.

There's a current discussion about this at
https://github.com/mozilla/source-map/issues/16, but it's unlikely there
will be progress on it since it's open since 2012 and last activity on
2015.

Anyway, I will be leaving this on hold until we have a solid solution to
work with this sort of section or when I need to provide source maps for
pure js files, which I am not been using too much lately :)