refactor: add project linting using xo code-style (#36)
We now are using xo instead of standard js since it is still almost
zero-config but allow for extra options. I was also quite sold on the
tabs vs spaces idea since those are dynamic representations of space and
can be tweaked to be displayed as any individual developer wants through
their text editor.

Additionally, we are also reformatting package.json and *.js files through
lint-staged on a precommit basis.